handshake working (#54)

+++ src/peer/types.rs
@@ -1,7 +1,6 @@
 #[derive(Debug)]
 pub struct Peer {
-    // TODO: move to peer module
     pub ip: String,
-    pub port: i64,
+    pub port: u16,
     pub peer_id: Vec<u8>,
 }
